refactor(tools): drop seven dormant paths across the tools layer
Landed:
- Delete the never-configured `maxSessions` capacity eviction from the direct
  Lean LSP adapter, along with the LRU picker and the session-freed waiter set
  it existed for. Idle eviction and the EMFILE recovery path are untouched.
  Three capacity-only test cases go with it; three that configured the cap
  inertly keep their assertions.
- Collapse the three single-caller delegation-availability wrappers and
  `annotateDelegationTool` into one exported `annotateDelegationAvailability`,
  and fold the four delegation test suites into one.
- Delete the external-inquiry execution mirror. Nothing has read
  `runs/<id>/ei/...` since #11050 removed the paths that named it, and
  `inquiry { command: 'read' }` already returns the full transcript.
- Delete the inline-comment runtime availability sentinel. The tool's `hosts:`
  block already keeps it off CLI and desktop rosters, so a missing provider is
  a startup bug and now throws instead of returning a plausible no-op.
- Fold `wolframScriptUtils` into `WolframTool` via the existing
  `runToolWithCheck`, dropping the `ExecResult` mirror type and the two null
  branches it forced.
- Replace the plan and todo no-work-plan arms, which returned a fabricated
  success, with a `ToolError`.
- Drop the never-written `toolchain` and never-read `pid` fields from the Lean
  server registry.

Skipped:
- Making `ToolCallContext.workPlanState` required. It deletes neither branch on
  its own (`getCurrentToolCallContext()` stays `| undefined`) and would force
  edits to five unrelated test files. Only the branch deletion landed.
- Step 3 of the inline-comment item (making `add()` non-null). The extension
  provider's null branch guards a controller cleared at deactivation.

knip-baseline rows removed:
- src/tools/delegation/delegationAvailability.ts / production-dead / exports /
  replaceDelegationDescriptionBlock
- src/tools/comment/InlineCommentTool.ts / production-dead / exports /
  InlineCommentInputSchema
